dom/security/test/mixedcontentblocker/test_bug1550792.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Bug 1550792: Block insecure subresource with non-https secure context parent</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>

  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>

<body>
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

let f = document.createElement("iframe");
f.src = "https://example.com/tests/dom/security/test/mixedcontentblocker/file_bug1550792.html";

window.addEventListener("message", (event) => {
  switch(event.data.type) {
    case 'http':
      is(event.data.status, "blocked", "nested load of http should be blocked.");
      break
    case 'https':
      is(event.data.status, "loaded", "nested load of https should not be blocked.");
      SimpleTest.finish();
      break;
  }
});

document.body.appendChild(f);

</script>
</body>
</html>
